feat: added authorization checks to Operate V1 API
- added permission checks to API endpoints
- added new exception and error handling for unauthorized requests

relates to https://github.com/camunda/camunda/issues/32408

+++ operate/webapp/src/main/java/io/camunda/operate/webapp/api/v1/rest/ProcessInstanceController.java
@@ -21,9 +21,13 @@ import io.camunda.operate.webapp.api.v1.entities.QueryValidator;
 import io.camunda.operate.webapp.api.v1.entities.Results;
 import io.camunda.operate.webapp.api.v1.entities.SequenceFlow;
 import io.camunda.operate.webapp.api.v1.exceptions.ClientException;
+import io.camunda.operate.webapp.api.v1.exceptions.ForbiddenException;
 import io.camunda.operate.webapp.api.v1.exceptions.ResourceNotFoundException;
 import io.camunda.operate.webapp.api.v1.exceptions.ServerException;
 import io.camunda.operate.webapp.api.v1.exceptions.ValidationException;
+import io.camunda.operate.webapp.security.permission.PermissionsService;
+import io.camunda.security.auth.Authorization;
+import io.camunda.zeebe.protocol.record.value.PermissionType;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
@@ -61,6 +65,7 @@ public class ProcessInstanceController extends ErrorController
   @Autowired private ProcessInstanceDao processInstanceDao;
   @Autowired private SequenceFlowDao sequenceFlowDao;
   @Autowired private FlowNodeStatisticsDao flowNodeStatisticsDao;
+  @Autowired private PermissionsService permissionsService;
 
   @Operation(
       summary = "Search process instances",
@@ -87,7 +92,14 @@ public class ProcessInstanceController extends ErrorController
             content =
                 @Content(
                     mediaType = MediaType.APPLICATION_PROBLEM_JSON_VALUE,
-                    schema = @Schema(implementation = Error.class)))
+                    schema = @Schema(implementation = Error.class))),
+        @ApiResponse(
+            description = ForbiddenException.TYPE,
+            responseCode = "403",
+            content =
+                @Content(
+                    mediaType = MediaType.APPLICATION_PROBLEM_JSON_VALUE,
+                    schema = @Schema(implementation = Error.class))),
       })
   @io.swagger.v3.oas.annotations.parameters.RequestBody(
       description = "Search process instances",
@@ -133,6 +145,7 @@ public class ProcessInstanceController extends ErrorController
     logger.debug("search for query {}", query);
     query = (query == null) ? new Query<>() : query;
     queryValidator.validate(query, ProcessInstance.class);
+    checkIdentityReadPermission();
     return processInstanceDao.search(query);
   }
 
@@ -156,6 +169,13 @@ public class ProcessInstanceController extends ErrorController
                     mediaType = MediaType.APPLICATION_PROBLEM_JSON_VALUE,
                     schema = @Schema(implementation = Error.class))),
         @ApiResponse(
+            description = ForbiddenException.TYPE,
+            responseCode = "403",
+            content =
+                @Content(
+                    mediaType = MediaType.APPLICATION_PROBLEM_JSON_VALUE,
+                    schema = @Schema(implementation = Error.class))),
+        @ApiResponse(
             description = ResourceNotFoundException.TYPE,
             responseCode = "404",
             content =
@@ -167,6 +187,7 @@ public class ProcessInstanceController extends ErrorController
   public ProcessInstance byKey(
       @Parameter(description = "Key of process instance", required = true) @PathVariable
           final Long key) {
+    checkIdentityReadPermission();
     return processInstanceDao.byKey(key);
   }
 
@@ -190,6 +211,13 @@ public class ProcessInstanceController extends ErrorController
                     mediaType = MediaType.APPLICATION_PROBLEM_JSON_VALUE,
                     schema = @Schema(implementation = Error.class))),
         @ApiResponse(
+            description = ForbiddenException.TYPE,
+            responseCode = "403",
+            content =
+                @Content(
+                    mediaType = MediaType.APPLICATION_PROBLEM_JSON_VALUE,
+                    schema = @Schema(implementation = Error.class))),
+        @ApiResponse(
             description = ResourceNotFoundException.TYPE,
             responseCode = "404",
             content =
@@ -204,7 +232,7 @@ public class ProcessInstanceController extends ErrorController
   public ChangeStatus delete(
       @Parameter(description = "Key of process instance", required = true) @Valid @PathVariable
           final Long key) {
-    // TODO: check for permissions https://github.com/camunda/camunda/issues/32408
+    checkIdentityDeletePermission();
     return processInstanceDao.delete(key);
   }
 
@@ -228,6 +256,13 @@ public class ProcessInstanceController extends ErrorController
                     mediaType = MediaType.APPLICATION_PROBLEM_JSON_VALUE,
                     schema = @Schema(implementation = Error.class))),
         @ApiResponse(
+            description = ForbiddenException.TYPE,
+            responseCode = "403",
+            content =
+                @Content(
+                    mediaType = MediaType.APPLICATION_PROBLEM_JSON_VALUE,
+                    schema = @Schema(implementation = Error.class))),
+        @ApiResponse(
             description = ResourceNotFoundException.TYPE,
             responseCode = "404",
             content =
@@ -239,7 +274,6 @@ public class ProcessInstanceController extends ErrorController
   public List<String> sequenceFlowsByKey(
       @Parameter(description = "Key of process instance", required = true) @PathVariable
           final Long key) {
-    processInstanceDao.byKey(key); // this is just to throw error if not found
     final Query<SequenceFlow> query =
         new Query<SequenceFlow>()
             .setFilter(new SequenceFlow().setProcessInstanceKey(key))
@@ -247,6 +281,8 @@ public class ProcessInstanceController extends ErrorController
     logger.debug("search for query {}", query);
     final QueryValidator<SequenceFlow> queryValidator = new QueryValidator<>();
     queryValidator.validate(query, SequenceFlow.class);
+    checkIdentityReadPermission();
+    processInstanceDao.byKey(key); // this is just to throw error if not found
     final Results<SequenceFlow> results = sequenceFlowDao.search(query);
     return results.getItems().stream()
         .map(SequenceFlow::getActivityId)
@@ -281,6 +317,13 @@ public class ProcessInstanceController extends ErrorController
                     mediaType = MediaType.APPLICATION_PROBLEM_JSON_VALUE,
                     schema = @Schema(implementation = Error.class))),
         @ApiResponse(
+            description = ForbiddenException.TYPE,
+            responseCode = "403",
+            content =
+                @Content(
+                    mediaType = MediaType.APPLICATION_PROBLEM_JSON_VALUE,
+                    schema = @Schema(implementation = Error.class))),
+        @ApiResponse(
             description = ResourceNotFoundException.TYPE,
             responseCode = "404",
             content =
@@ -292,7 +335,22 @@ public class ProcessInstanceController extends ErrorController
   public Collection<FlowNodeStatistics> getStatistics(
       @Parameter(description = "Key of process instance", required = true) @PathVariable
           final Long key) {
+    checkIdentityReadPermission();
     processInstanceDao.byKey(key); // this is just to throw error if not found
     return flowNodeStatisticsDao.getFlowNodeStatisticsForProcessInstance(key);
+  }
+
+  private void checkIdentityReadPermission() {
+    if (!permissionsService.hasPermissionForProcess(
+        Authorization.WILDCARD, PermissionType.READ_PROCESS_INSTANCE)) {
+      throw new ForbiddenException("No read permission for process instances");
+    }
+  }
+
+  private void checkIdentityDeletePermission() {
+    if (!permissionsService.hasPermissionForProcess(
+        Authorization.WILDCARD, PermissionType.DELETE_PROCESS_INSTANCE)) {
+      throw new ForbiddenException("No delete permission for process instances");
+    }
   }
 }
